Moved `db` "models" (structs) to own module Should make project structure a bit cleaner. Also, the db models module re-exports the models from the initial schema.rs definition.
