+++ src/db/migration/mod.rs
@@ -1,0 +1,2 @@
+mod migration;
+pub use migration::*;
