+++ src/db/mod.rs
@@ -1,9 +1,10 @@
-pub mod attachments;
 mod migration;
+mod models;
+pub use migration::*;
+
+pub mod attachments;
 pub mod reviews;
 pub mod submissions;
 pub mod todos;
 pub mod users;
 pub mod workshops;
-
-pub use migration::*;
