+++ src/db/models/mod.rs
@@ -1,0 +1,3 @@
+mod models;
+pub use crate::models::*;
+pub use models::*;
