+++ src/utils/mod.rs
@@ -1,4 +1,3 @@
 pub mod json;
 mod path;
-
 pub use path::*;
